Simple linear regression.
In this lab, we learn how to use the scikit-learn library to implement Simple linear regression. We download a dataset that is related to fuel consumption and Carbon dioxide emission of cars. Then, we split our data into training and test sets, create a model using a training set, evaluate your model using the test set, and finally use the model to predict unknown value.